ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Embeddings
sentence-transformers==3.3.0
torch==2.5.0

# Document extraction
PyMuPDF==1.25.0
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.5
pandas==2.2.0

# OCR
pytesseract==0.3.13
Pillow==11.0.0

# Text processing
langchain-text-splitters==0.3.0
langdetect==1.0.9

# Vector store
qdrant-client==1.12.0

# LLM
anthropic>=0.42.0

# Discovery APIs
arxiv==2.1.3
httpx==0.28.0

# Storage
boto3==1.35.0

# Utilities
python-multipart==0.0.12
